chore(Seeds): rename seed
